==> packages/exploratory-data-analysis/encuestas/02-eda-encuesta_ingresantes_2024.ipynb <==
import pandas as pd
import numpy as np

# Encuestas a ingresantes

A continuación vamos a estudiar los datos que nos pasaron de encuestas que se le realizaron a ingresantes de la FCEN.

Los datasets originales NO serán subidos al repositorio, pues veremos que contienen información delicada.

La información se corresponde con datos del segundo cuatrimestre de 2024 en adelante.

## RIESGO +ACOMPAÑAMIENTO 1C2024

# Leer el archivo Excel
xls = pd.ExcelFile("../../../assets/ENCUESTAS+LISTADOS/RIESGO +ACOMPAÑAMIENTO 1C2024.xlsx")
hojas = xls.sheet_names
print(hojas)

### Primer hoja

primer_c_2024_1 = pd.read_excel(xls, sheet_name=hojas[0])

primer_c_2024_1.columns[primer_c_2024_1.columns == 'Puntaje_riesgo']

primer_c_2024_1['Puntaje_riesgo'].unique()

El único valor faltante se debe a que respondió una de las opciones que no estaba escrita exactamente igual en la tabla de puntajes

print(primer_c_2024_1['Puntaje_riesgo'].isnull().sum())

primer_c_2024_1.loc[primer_c_2024_1['Puntaje_riesgo'].isnull(), 'Puntaje_riesgo']

primer_c_2024_1.shape

primer_c_2024_1['Nº de documento'].nunique()

frecuencia = primer_c_2024_1['Nº de documento'].value_counts()
# Muestra los valores que aparecen más de una vez
print("Frecuencia de valores en 'Id':")
with pd.option_context('display.max_rows', None):
    print(frecuencia[frecuencia > 1])

primer_c_2024_1[primer_c_2024_1['Nº de documento'].isin(frecuencia[frecuencia > 1].index)].sort_values('Nº de documento').to_csv('duplicados_2024.csv', index=False)

# Convertir la columna a string (por si acaso no lo está)
primer_c_2024_1['Nº de documento'] = primer_c_2024_1['Nº de documento'].astype(str)

# Eliminar espacios al principio y al final
primer_c_2024_1['Nº de documento'] = primer_c_2024_1['Nº de documento'].str.strip()

primer_c_2024_1['Nivel'].unique()

print(primer_c_2024_1['Nivel'].isnull().sum())

### Segunda hoja

Pareciera tener exactamente el mismo formato que la del año 2023.

Hay columnas que fueron autocompletadas con formulas de excel:

* Id columna: concatenación entre "Columna" y "Valor"
* Valor: Toma "valor=puntaje" y se queda con lo que está antes del símbolo =
* Puntaje: Toma "valor=puntaje" y se queda con lo que está despues del símbolo =

Despues me va a interesar ver si contienen las mismas preguntas

primer_c_2024_2 = pd.read_excel(xls, sheet_name=hojas[1], header=1)
primer_c_2024_2 = primer_c_2024_2[:152]
primer_c_2024_2

### Tercer hoja
Usaron la misma grilla que en el año 2023.

primer_c_2024_3 = pd.read_excel(xls, sheet_name=hojas[2])
primer_c_2024_3

### Cuarta hoja

Como mencioné para 2023, pensaría que a partir de esta tabla se creó la que se tiene en la primer hoja del excel.

primer_c_2024_4 = pd.read_excel(xls, sheet_name=hojas[3], header=1)

primer_c_2024_4.shape

primer_c_2024_4['Nº de documento'].isnull().sum()

primer_c_2024_4['Nº de documento'].nunique()

primer_c_2024_1['Nº de documento'].nunique()

primer_c_2024_1['Nº de documento'].isnull().sum()

# Convertir la columna a string (por si acaso no lo está)
primer_c_2024_4['Nº de documento'] = primer_c_2024_4['Nº de documento'].astype(str)

# Eliminar espacios al principio y al final
primer_c_2024_4['Nº de documento'] = primer_c_2024_4['Nº de documento'].str.strip()
# Encuentra los valores que están en ambas columnas
# Convierte la columna a cadenas de texto
intersection = primer_c_2024_1[primer_c_2024_1['Nº de documento'].isin(primer_c_2024_4['Nº de documento'])]['Nº de documento']

# Elimina duplicados (si es necesario)
intersection = intersection.drop_duplicates()

print("Intersección:", intersection.tolist())

# Encuentra los valores que están en ambas columnas
# Convierte la columna a cadenas de texto
intersection = primer_c_2024_4[primer_c_2024_4['Nº de documento'].isin(primer_c_2024_1['Nº de documento'])]['Nº de documento']

# Elimina duplicados (si es necesario)
intersection = intersection.drop_duplicates()

print("Intersección:", intersection.tolist())

## Informe final +Acompañamiento-2C2024 ALGEBRA.xlsx

# Leer el archivo Excel
xls_2C = pd.ExcelFile("../../../assets/ENCUESTAS+LISTADOS/Informe final +Acompañamiento-2C2024 ALGEBRA.xlsx")
hojas_2C = xls_2C.sheet_names
print(hojas_2C)

segundo_c_2024 = pd.read_excel(xls_2C, sheet_name=hojas_2C[0])

En las últimas filas del documento se había armado una tabla a modo de resumen. Elimino dichas filas para poder hacer un análisis de datos.

segundo_c_2024 = segundo_c_2024[:392]

segundo_c_2024.to_csv('../../../assets/InformeFinal_2C2024/algebra.csv', index=False)

segundo_c_2024_csv = pd.read_csv('../../../assets/InformeFinal_2C2024/algebra.csv')

segundo_c_2024_csv.shape

segundo_c_2024_csv['Apellido'] = segundo_c_2024_csv['Apellido'].str.strip()
segundo_c_2024_csv['Nombre'] = segundo_c_2024_csv['Nombre'].str.strip()

# Crear una columna combinada de "Apellido" y "Nombre"
segundo_c_2024_csv['Combinado'] = segundo_c_2024_csv['Apellido'] + ' ' + segundo_c_2024_csv['Nombre']

# Contar la cantidad de pares únicos de "Apellido" y "Nombre"
pares_unicos = segundo_c_2024_csv['Combinado'].nunique()

print(f"La cantidad de pares únicos de 'Apellido' y 'Nombre' es: {pares_unicos}")

## Estudiantes, notas, resultados F1 2C2024.xlsx

# Leer el archivo Excel
xls_2C_fisica = pd.ExcelFile("../../../assets/ENCUESTAS+LISTADOS/Estudiantes, notas, resultados F1 2C2024.xlsx")
hojas_2C_fisica = xls_2C_fisica.sheet_names
print(hojas_2C_fisica)

## Semáforo 2C 2024.xlsx

# Leer el archivo Excel
xls_2C_2024 = pd.ExcelFile("../../../assets/ENCUESTAS+LISTADOS/Semáforo 2C 2024.xlsx")
hojas_2C_2024 = xls_2C_2024.sheet_names
print(hojas_2C_2024)

segundo_c_2024_1 = pd.read_excel(xls_2C_2024, sheet_name=hojas_2C_2024[0])

segundo_c_2024_1.columns

segundo_c_2024_1['Nº de documento'].nunique()

segundo_c_2024_1['Nº de documento'].shape

frecuencia = segundo_c_2024_1['Nº de documento'].value_counts()
# Muestra los valores que aparecen más de una vez
print("Frecuencia de valores en 'Id':")
with pd.option_context('display.max_rows', None):
    print(frecuencia[frecuencia > 1])

segundo_c_2024_1
