==> packages/exploratory-data-analysis/encuestas/03-eda-encuestas_ingresantes_datos_utilizados.ipynb <==
import sys
sys.path.insert(0, '../')

import pandas as pd
import numpy as np
import helpers_encuestas
import helpers_eda_inicial

# Leer el archivo Excel
encuesta_2c_2023 = df = pd.read_csv('../../../assets/silver/RIESGO +ACOMPAÑAMIENTO 2C 2023/RIESGO +ACOMPAÑAMIENTO 2C 2023.xlsx - Encuesta a Estudiantes Iniciale.csv')
xls = pd.ExcelFile("../../../assets/bronze/ENCUESTAS+LISTADOS/RIESGO +ACOMPAÑAMIENTO 1C2024.xlsx")
primer_c_2024_1 = pd.read_excel(xls, sheet_name='Encuesta a Estudiantes Iniciale')
xls_2C_2024 = pd.ExcelFile("../../../assets/bronze/ENCUESTAS+LISTADOS/Semáforo 2C 2024.xlsx")
segundo_c_2024_1 = pd.read_excel(xls_2C_2024, sheet_name='Encuesta a Ingresantes de EXACT')

# listas de columnas
cols_encuesta_2023 = encuesta_2c_2023.columns.to_list()
cols_primer_2024_1 = primer_c_2024_1.columns.to_list()
cols_segundo_2024_1 = segundo_c_2024_1.columns.to_list()

# Lo metemos todo en un diccionario
columnas_dict = {
    'encuesta_2c_2023': cols_encuesta_2023,
    'primer_c_2024_1': cols_primer_2024_1,
    'segundo_c_2024_1': cols_segundo_2024_1
}

# Creamos el DataFrame, pandas se encarga de alinear y rellenar con NaN donde falte
df_columnas = pd.DataFrame(dict([(k, pd.Series(v)) for k, v in columnas_dict.items()]))

print(df_columnas)

Se analizó manualmente las preguntas de cada año, se encontró que los primeros dos cuatrimestres se utilizó el mismo conjunto de preguntas. Sin embargo, en el segundo cuatrimestre de 2024 se cambiaron algunas de las preguntas. En algunos casos solo cambio el formato de la pregunta, en otro se agregaron opciones, etc. Debemos tratar cada uno de estos casos para poner unificar los datasets.

encuesta_2c_2023['¿Cuándo ingresaste a Exactas? '].unique()

primer_c_2024_1['¿Cuándo ingresaste a Exactas? '].unique()

print(segundo_c_2024_1.columns.to_list())

segundo_c_2024_1['¿En qué año ingresaste a Exactas?'].unique()

encuesta_2c_2023['¿Cuándo ingresaste a Exactas?  [Otro]'].unique()

primer_c_2024_1['¿Cuándo ingresaste a Exactas?  [Otro]'].unique()

Pensé que las respuestas más "raras" podían llegar a coincidir porque se ingresó 2 veces la misma encuesta.

np.isin(encuesta_2c_2023['Nº de documento'].astype(str).unique(), primer_c_2024_1['Nº de documento'].astype(str).unique()).sum()

encuesta_2c_2023['Nº de documento']

primer_c_2024_1['Nº de documento']

np.isin(primer_c_2024_1['Nº de documento'].unique(), encuesta_2c_2023['Nº de documento'].unique()).sum()

primer_c_2024_1[primer_c_2024_1['Nº de documento'].isin(encuesta_2c_2023['Nº de documento'].unique())].to_csv('repetidos.csv', index=False)

encuesta_2c_2023[encuesta_2c_2023['Nº de documento'].isin(primer_c_2024_1['Nº de documento'].unique())].to_csv('repetidos_2.csv', index=False)

Vemos exactamente las filas con esa respuesta:

primer_c_2024_1[primer_c_2024_1['¿Cuándo ingresaste a Exactas?  [Otro]'] == 'Ingrese a fisica en 2020. Me cambie a biologia en 2023']

encuesta_2c_2023[encuesta_2c_2023['¿Cuándo ingresaste a Exactas?  [Otro]'] == 'Ingrese a fisica en 2020. Me cambie a biologia en 2023']

## Prueba con otro de los datasets

primer_c_2024_4 = pd.read_excel(xls, sheet_name='Copia de Encuesta a Ingresantes', header=1)

np.isin(encuesta_2c_2023['Nº de documento'].astype(str).unique(), primer_c_2024_4['Nº de documento'].astype(str).unique()).sum()

primer_c_2024_4['Nº de documento'].isna().sum()

primer_c_2024_4['Nº de documento'].nunique()

primer_c_2024_4.shape

## Hojas con las que trabajamos

Pareciera que en realidad la hoja correcta es la que se llama 'Copia de Encuesta a Ingresantes'. Por ahora voy a trabajar con dicha tabla y volveré a consultar. El problema con esta hoja es que no tiene asignada la columna con los puntos, por lo tanto, vale la pena consultar si tienen otra versión donde sí asignaron valores al semáforo.

* encuesta_2c_2023

* primer_c_2024_4

* segundo_c_2024_1

# Leer el archivo Excel
encuesta_2c_2023 = df = pd.read_csv('../../../assets/silver/RIESGO +ACOMPAÑAMIENTO 2C 2023/RIESGO +ACOMPAÑAMIENTO 2C 2023.xlsx - Encuesta a Estudiantes Iniciale.csv')
xls = pd.ExcelFile("../../../assets/bronze/ENCUESTAS+LISTADOS/RIESGO +ACOMPAÑAMIENTO 1C2024.xlsx")
primer_c_2024_4 = pd.read_excel(xls, sheet_name='Copia de Encuesta a Ingresantes', header=1)
xls_2C_2024 = pd.ExcelFile("../../../assets/bronze/ENCUESTAS+LISTADOS/Semáforo 2C 2024.xlsx")
segundo_c_2024_1 = pd.read_excel(xls_2C_2024, sheet_name='Encuesta a Ingresantes de EXACT')

print(primer_c_2024_1.shape)
print(primer_c_2024_1['Nº de documento'].isna().sum())
print(primer_c_2024_1['Nº de documento'].nunique())

np.isin(encuesta_2c_2023['Nº de documento'].astype(str).unique(), primer_c_2024_1['Nº de documento'].astype(str).unique()).sum()

np.isin(primer_c_2024_1['Nº de documento'].astype(str).unique(), primer_c_2024_4['Nº de documento'].astype(str).unique()).sum()

np.isin(primer_c_2024_4['Nº de documento'].astype(str).unique(), segundo_c_2024_1['Nº de documento'].astype(str).unique()).sum()

np.isin(primer_c_2024_1['Nº de documento'].astype(str).unique(), segundo_c_2024_1['Nº de documento'].astype(str).unique()).sum()

np.isin(encuesta_2c_2023['Nº de documento'].astype(str).unique(), segundo_c_2024_1['Nº de documento'].astype(str).unique()).sum()

Primero vemos la cantidad de nulos, pero para el trabajo solo podemos trabajar con aquellos que tienen el DNI ingresado

print(encuesta_2c_2023.shape)
print(encuesta_2c_2023['Nº de documento'].isna().sum())
print(encuesta_2c_2023['Nº de documento'].nunique())

print(primer_c_2024_4.shape)
print(primer_c_2024_4['Nº de documento'].isna().sum())
print(primer_c_2024_4['Nº de documento'].nunique())

print(segundo_c_2024_1.shape)
print(segundo_c_2024_1['Nº de documento'].isna().sum())
print(segundo_c_2024_1['Nº de documento'].nunique())

#Eliminamos las columnas vacías y las que sabemos que no usaremos
encuesta_2c_2023.drop(columns=['Libreta Universitaria Nº', 'Apellido/s y Nombres', 'Lenguaje inicial', 'Semilla', 'Contraseña', 'Fecha de envío', 'Última página', 'Comentarios Finales'], inplace=True)
primer_c_2024_4.drop(columns=['Apellido/s y Nombres', 'Lenguaje inicial', 'Semilla', 'Fecha de envío', 'Última página', 'Unnamed: 5', 'Unnamed: 8', 'Unnamed: 65', 'Unnamed: 66', 'Comentarios Finales'], inplace = True)
segundo_c_2024_1.drop(columns=['Apellido/s y Nombres', 'Lenguaje inicial', 'Semilla', 'Fecha de envío', 'Última página','Unnamed: 5', 'Unnamed: 8', 'Unnamed: 65', 'Unnamed: 66', 'Comentarios Finales'], inplace=True)

# saco los valores nulos
encuesta_2c_2023.dropna(subset=['Nº de documento'], inplace=True)
primer_c_2024_4.dropna(subset=['Nº de documento'], inplace=True)
segundo_c_2024_1.dropna(subset=['Nº de documento'], inplace=True)

# Saco los duplicados
encuesta_2c_2023.drop_duplicates(inplace=True)
primer_c_2024_4.drop_duplicates(inplace=True)
segundo_c_2024_1.drop_duplicates(inplace=True)

segundo_c_2024_1['Nº de documento'].value_counts()

dni_repetido = segundo_c_2024_1['Nº de documento'].value_counts()[segundo_c_2024_1['Nº de documento'].value_counts() == 2].index.tolist()

sin_id = segundo_c_2024_1.drop(columns=['ID de respuesta'], inplace=False)
sin_id.drop_duplicates(inplace=True)

sin_id[sin_id['Nº de documento'].isin(dni_repetido)]

df_filtrado = sin_id[sin_id['Nº de documento'].isin(dni_repetido)]
fila1 = df_filtrado.iloc[0]
fila2 = df_filtrado.iloc[1]

columnas_diferentes = []

for col in sin_id.columns:
    val1, val2 = fila1[col], fila2[col]
    # Si ambos son NaN → continuar
    if pd.isna(val1) and pd.isna(val2):
        continue
    # Si son distintos (incluyendo si uno es NaN y el otro no)
    if val1 != val2:
        columnas_diferentes.append(col)

print(columnas_diferentes)

Se ve que es una persona que respondió más de una vez la encuesta y sus respuestas varian levemente, nos quedamos con una de las filas.

Para los otros años me dijeron que me podía quedar con cualquiera de las filas, de manera indistinta.

segundo_c_2024_1 = segundo_c_2024_1.drop_duplicates(subset='Nº de documento', keep='last')
encuesta_2c_2023 = encuesta_2c_2023.drop_duplicates(subset='Nº de documento', keep='last')
primer_c_2024_4 = primer_c_2024_4.drop_duplicates(subset='Nº de documento', keep='last')

## Comienzo a arreglar columnas

Algunas columnas de segundo_c_2024_1 se llaman distinto porque cambiaron la forma de preguntar pero significan lo mismo. Lo miramos

segundo_c_2024_1['¿En qué año ingresaste a Exactas?'].unique()

segundo_c_2024_1['¿En qué cuatrimestre ingresaste?'].unique()

encuesta_2c_2023['¿Cuándo ingresaste a Exactas? '].unique()

encuesta_2c_2023['¿Cuándo ingresaste a Exactas?  [Otro]'].unique()

primer_c_2024_4['¿Cuándo ingresaste a Exactas? '].unique()

primer_c_2024_4['¿Cuándo ingresaste a Exactas?  [Otro]'].unique()

helpers_eda_inicial.dataset_profiling(primer_c_2024_4, name='1c2024')

helpers_eda_inicial.dataset_profiling(encuesta_2c_2023, name='2c2023')
helpers_eda_inicial.dataset_profiling(segundo_c_2024_1, name='2c2024')
