==> packages/exploratory-data-analysis/encuestas/04-eda-encuestras_columnas_iguales.ipynb <==
import sys
sys.path.insert(0, '../')

import pandas as pd
import numpy as np

import helpers_cbc
import helpers_encuestas

Se encontraron multiples inconsistencias entre los archivos de las encuestas proporcionados originalmente, por lo que se nos dio acceso a los resultados de las encuestas orinales. Esto implica que no tienen ningun procesamiento previo.

xls_2c2023_v2= pd.ExcelFile("../../../assets/bronze/Encuestas_v2/ingresantes 2c 2023.xlsx") 
encuesta_2c_2023_v2 = pd.read_excel(xls_2c2023_v2)
xls_v2 = pd.ExcelFile("../../../assets/bronze/Encuestas_v2/ingresantes 1c 2024.xlsx") 
primer_c_2024_1_v2 = pd.read_excel(xls_v2)
xls_2C_2024_v2 = pd.ExcelFile("../../../assets/bronze/Encuestas_v2/ingresantes 2c 2024.xlsx")
segundo_c_2024_1_v2 = pd.read_excel(xls_2C_2024_v2)

np.isin(encuesta_2c_2023_v2['Nº de documento'].astype(str).unique(), primer_c_2024_1_v2['Nº de documento'].astype(str).unique()).sum()

np.isin(encuesta_2c_2023_v2['Nº de documento'].astype(str).unique(), segundo_c_2024_1_v2['Nº de documento'].astype(str).unique()).sum()

np.isin(primer_c_2024_1_v2['Nº de documento'].astype(str).unique(), segundo_c_2024_1_v2['Nº de documento'].astype(str).unique()).sum()

Ahora solo tinen un DNI en común. Sospecho que lo que tienen en común son los NA

encuesta_2c_2023 = df = pd.read_csv('../../../assets/silver/RIESGO +ACOMPAÑAMIENTO 2C 2023/RIESGO +ACOMPAÑAMIENTO 2C 2023.xlsx - Encuesta a Estudiantes Iniciale.csv')
xls = pd.ExcelFile("../../../assets/bronze/ENCUESTAS+LISTADOS/RIESGO +ACOMPAÑAMIENTO 1C2024.xlsx")
primer_c_2024_4 = pd.read_excel(xls, sheet_name='Copia de Encuesta a Ingresantes', header=1)
xls_2C_2024 = pd.ExcelFile("../../../assets/bronze/ENCUESTAS+LISTADOS/Semáforo 2C 2024.xlsx")
segundo_c_2024_1 = pd.read_excel(xls_2C_2024, sheet_name='Encuesta a Ingresantes de EXACT')

np.isin(encuesta_2c_2023['Nº de documento'].astype(str).unique(), encuesta_2c_2023_v2['Nº de documento'].astype(str).unique()).sum()

print(encuesta_2c_2023.shape)
print(encuesta_2c_2023['Nº de documento'].isna().sum())
print(encuesta_2c_2023['Nº de documento'].nunique())

print(encuesta_2c_2023_v2.shape)
print(encuesta_2c_2023_v2['Nº de documento'].isna().sum())
print(encuesta_2c_2023_v2['Nº de documento'].nunique())

# Limpio un poco las encuestas

helpers_encuestas.delete_null_values(encuesta_2c_2023_v2, primer_c_2024_1_v2, segundo_c_2024_1_v2)

encuesta_2c_2023_v2.to_csv('../../../assets/silver/encuestas_con_dni/encuesta_2c_2023.csv', index=False)

primer_c_2024_1_v2.to_csv('../../../assets/silver/encuestas_con_dni/encuesta_primer_c_2024.csv', index=False)

segundo_c_2024_1_v2.to_csv('../../../assets/silver/encuestas_con_dni/encuesta_segundo_c_2024.csv', index=False)

# Unifico las tablas con las preguntas que son iguales

segundo_c_2023_limpia = pd.read_csv('../../../assets/silver/encuestas_limpias/encuesta_2c_2023 - encuesta_2c_2023.csv')
primer_c_2024_limpia = pd.read_csv('../../../assets/silver/encuestas_limpias/encuesta_primer_c_2024 - encuesta_primer_c_2024.csv')
segundo_c_2024_limpia = pd.read_csv('../../../assets/silver/encuestas_limpias/encuesta_segundo_c_2024 - encuesta_segundo_c_2024.csv')

# listas de columnas
cols_encuesta_2023 = segundo_c_2023_limpia.columns.to_list()
cols_primer_2024_1 = primer_c_2024_limpia.columns.to_list()
cols_segundo_2024_1 = segundo_c_2024_limpia.columns.to_list()

# Lo metemos todo en un diccionario
columnas_dict = {
    'encuesta_2c_2023': cols_encuesta_2023,
    'primer_c_2024_1': cols_primer_2024_1,
    'segundo_c_2024_1': cols_segundo_2024_1
}

# Creamos el DataFrame, pandas se encarga de alinear y rellenar con NaN donde falte
df_columnas = pd.DataFrame(dict([(k, pd.Series(v)) for k, v in columnas_dict.items()]))

print(df_columnas)

df_columnas.to_csv('../../../assets/silver/encuestas_con_dni/columnas_encuestas.csv', index=False)

me quedo solamente con las que tienen en común.

segundo_c_2023_limpia.drop_duplicates(subset=['Nº de documento'], keep='last', inplace=True)
primer_c_2024_limpia.drop_duplicates(subset=['Nº de documento'], keep='last', inplace=True)
segundo_c_2024_limpia.drop_duplicates(subset=['Nº de documento'], keep='last', inplace=True)

segundo_c_2023_limpia.to_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_2c_2023_limpia.csv', index=False)
primer_c_2024_limpia.to_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_primer_c_2024_limpia.csv', index=False)
segundo_c_2024_limpia.to_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_segundo_c_2024_limpia.csv', index=False)

segundo_c_2023_limpia['Nº de documento'].nunique(), primer_c_2024_limpia['Nº de documento'].nunique(), segundo_c_2024_limpia['Nº de documento'].nunique()

# Supongamos que ya tenés los 3 DataFrames cargados:
# segundo_c_2023_limpia, primer_c_2024_limpia, segundo_c_2024_limpia

# Encontramos las columnas comunes
columnas_comunes = list(
    set(segundo_c_2023_limpia.columns) &
    set(primer_c_2024_limpia.columns) &
    set(segundo_c_2024_limpia.columns)
)

# Nos quedamos con las columnas comunes en cada tabla
df_2023 = segundo_c_2023_limpia[columnas_comunes]
df_2024_1 = primer_c_2024_limpia[columnas_comunes]
df_2024_2 = segundo_c_2024_limpia[columnas_comunes]

# Concatenamos las tablas
df_unido = pd.concat([df_2023, df_2024_1, df_2024_2], ignore_index=True)


df_unido.columns

for col in df_unido.columns:
    print(f"- `{col}`")

df_unido['Nº de documento'].nunique()

df_unido.shape

df_unido = df_unido.rename(columns={'Nº de documento': 'dni'})

df_unido.to_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_unida.csv', index=False)

df_unido = pd.read_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_unida.csv')

CBC = '../../../assets/bronze/CBC/Sitacad_Tesis (1).xlsx'
df_calificaciones, df_carreras = helpers_cbc.get_data(CBC)

df_calificaciones[df_calificaciones['Dni'].isin(df_unido['dni'])]['Dni'].nunique()

df_unido[~df_unido['dni'].isin(df_calificaciones['Dni'])]['¿En que año iniciaste el CBC para Exactas?'].value_counts()

df_unido[~df_unido['dni'].isin(df_calificaciones['Dni'])][df_unido['¿En que año iniciaste el CBC para Exactas?'] == '2022']['dni']
