==> packages/exploratory-data-analysis/encuestas/05-encuestas_unificacion.ipynb <==
import sys
sys.path.insert(0, '../')

import pandas as pd
import numpy as np

import helpers_cbc
import helpers_encuestas

encuesta_2c_2023_v2 = pd.read_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_2c_2023_limpia.csv')
primer_c_2024_1_v2 = pd.read_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_primer_c_2024_limpia.csv')
segundo_c_2024_1_v2 = pd.read_csv('../../../assets/silver/encuestas_limpias_sin_duplicados/encuesta_segundo_c_2024_limpia.csv')

Ya tengo las columnas donde cambiaron minimamente la pregunta, quiero unificar el nombre de las mismas para generar un dataset unico

# Saco los valores nulos
helpers_encuestas.delete_null_values(encuesta_2c_2023_v2, primer_c_2024_1_v2, segundo_c_2024_1_v2)
helpers_encuestas.drop_dni_duplicates(encuesta_2c_2023_v2, primer_c_2024_1_v2, segundo_c_2024_1_v2)

encuesta_2c_2023_v2['Nº de documento'].nunique(), primer_c_2024_1_v2['Nº de documento'].nunique(), segundo_c_2024_1_v2['Nº de documento'].nunique()

Vemos los tipos asignados

dtypes = encuesta_2c_2023_v2.dtypes.to_frame(name="Data Type")
print(dtypes.to_markdown())

dtypes = primer_c_2024_1_v2.dtypes.to_frame(name="Data Type")
print(dtypes.to_markdown())

dtypes = segundo_c_2024_1_v2.dtypes.to_frame(name="Data Type")
print(dtypes.to_markdown())

print(primer_c_2024_1_v2['¿Cómo evaluás este espacio destinado al estudio?'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

print(segundo_c_2024_1_v2['¿Cómo evaluás este espacio destinado al estudio?'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

print(encuesta_2c_2023_v2['¿Contas con un espacio para el estudio en tu hogar, qué características tiene? [Es adecuado]'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

print(encuesta_2c_2023_v2['¿Contas con un espacio para el estudio en tu hogar, qué características tiene? [Es medianamente adecuado]'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

print(encuesta_2c_2023_v2['¿Contas con un espacio para el estudio en tu hogar, qué características tiene? [Es inadecuado]'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

Cambio mucho la pregunta y no tengo forma de unificarla entre los 3 cuatrimestres

helpers_encuestas.drop_columns_final_version(encuesta_2c_2023_v2, primer_c_2024_1_v2, segundo_c_2024_1_v2)

print(encuesta_2c_2023_v2['¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Análisis Matemático]'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

print(primer_c_2024_1_v2['Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento Científico (IPC)]'].value_counts().to_frame(name="Cantidad").reset_index().rename(columns={"index":"Respuesta"}).to_markdown())

columnas_a_transformar = ['Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento Científico (IPC)]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento de la Sociedad y el Estado (ICSE)]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Análisis Matemático]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Algebra]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Matemática]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Física]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Química]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Biología]']

for col in columnas_a_transformar:
    print(primer_c_2024_1_v2[col].unique())
    print(segundo_c_2024_1_v2[col].unique())

def transformar_dificultad(df, columna_original):

    # Creamos un diccionario con el mapeo deseado
    mapeo_dificultad = {
        'Baja': 'No',
        'Muy Baja': 'No',
        'Media': 'No',
        'Alta': 'Si',
        'Muy Alta': 'Si',
        'No la cursé': np.nan
    }

    # Aplicamos la transformación
    df[columna_original] = df[columna_original].replace(mapeo_dificultad)

columnas_a_transformar = ['Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento Científico (IPC)]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento de la Sociedad y el Estado (ICSE)]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Análisis Matemático]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Algebra]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Matemática]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Física]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Química]',
                        'Indicanos el nivel de dificultad que te generaron las materias del CBC [Biología]']
for columna in columnas_a_transformar:
    transformar_dificultad(primer_c_2024_1_v2, columna)
    transformar_dificultad(segundo_c_2024_1_v2, columna)

# Diccionario de renombrado
renombrar_columnas = {
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento Científico (IPC)]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Introducción al Pensamiento Científico (IPC)]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Int. al Conocimiento de la Sociedad y el Estado (ICSE)]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Introducción al Conocimiento de la Sociedad y el Estado (ICSE)]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Análisis Matemático]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Análisis Matemático]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Algebra]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Algebra]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Matemática]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Matemática]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Física]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Fisica]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Química]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Quimica]',
    'Indicanos el nivel de dificultad que te generaron las materias del CBC [Biología]': '¿Cuál o cuáles materias del CBC te generaron mayor dificultad? [Biología]'
}

# Renombrar las columnas
primer_c_2024_1_v2.rename(columns=renombrar_columnas, inplace=True)
segundo_c_2024_1_v2.rename(columns=renombrar_columnas, inplace=True)

# Diccionario de renombrado
renombrar_columnas = {
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Tutores del CBC]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Tutores]',
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Compañeras/os de estudio]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Compañeras/os de estudio]',
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Profesor/es particular/es]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Profesor/es particular/es]',
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Familiares]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Familiares]',
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Amigos]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Amigos]',
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Lo resolví por mi cuenta]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Lo resolví por mi cuenta]',
    '¿En caso de haber tenido dificultades, cómo las resolviste? ¿recurriste a alguien? [Otro]': '¿Cómo resolviste estas dificultades, recurriste a alguien? [Otro]',
    'Indicá la carrera y en que universidad cursaste\xa0': 'Indicá la carrera y en que universidad cursaste',
    '¿Cual es la ocupación del/a principal sosten de tu hogar? [Otro]': '¿Cual es la ocupacion del/a principal sosten de tu hogar? [Otro]',
    '¿Cual es la ocupación del/a principal sosten de tu hogar?': '¿Cual es la ocupacion del/a principal sosten de tu hogar?'
}

# Renombrar las columnas
primer_c_2024_1_v2.rename(columns=renombrar_columnas, inplace=True)
segundo_c_2024_1_v2.rename(columns=renombrar_columnas, inplace=True)

Me fijo que todos quedaran con los mismos nombres de columnas

# Obtener las columnas de cada dataset como sets
cols1 = set(encuesta_2c_2023_v2.columns)
cols2 = set(primer_c_2024_1_v2.columns)
cols3 = set(segundo_c_2024_1_v2.columns)

# Ver si todos tienen las mismas columnas
if cols1 == cols2 == cols3:
    print("Todos los datasets tienen las mismas columnas.")
else:
    print("Hay diferencias en las columnas:")
    print("Columnas en df1 pero no en df2:", cols1 - cols2)
    print("Columnas en df2 pero no en df1:", cols2 - cols1)
    print("Columnas en df1 pero no en df3:", cols1 - cols3)
    print("Columnas en df3 pero no en df1:", cols3 - cols1)
    print("Columnas en df2 pero no en df3:", cols2 - cols3)
    print("Columnas en df3 pero no en df2:", cols3 - cols2)

encuesta_2c_2023_v2["origen"] = "2c2023"
primer_c_2024_1_v2["origen"] = "1c2024"
segundo_c_2024_1_v2["origen"] = "2c2024"

# Unir los tres datasets
df_concatenado = pd.concat([encuesta_2c_2023_v2, primer_c_2024_1_v2, segundo_c_2024_1_v2], ignore_index=True)

# Verificá el resultado
print(df_concatenado.shape)
df_concatenado.head()

df_concatenado.dtypes

df_concatenado.to_csv('../../../assets/silver/encuestas_unificadas/encuestas_concatenadas.csv', index=False)
